refactor: rename inference preparation module
