+++ crates/goose/src/agents/state_machine/inference_preparation.rs
