+++ crates/goose/src/agents/state_machine/mod.rs
@@ -6,12 +6,12 @@
 //! configuration is part of `Agent::reply`, not the state-machine protocol.
 
 mod effects;
+mod inference_preparation;
 mod ops_bang_shell;
 mod ops_compaction;
 mod ops_doctor;
 mod ops_entry_hook;
 mod ops_exit_on_error;
-mod ops_inference_request_preparer;
 mod ops_llm;
 mod ops_maxturns;
 mod ops_project;
